Initial commit
Well, it works, but it's nasty-looking, and a lot of the things are
hard-coded, so needs a deep-clean! Also, I'll strive to convert parts
of this into a cargo-plugin for easier/cleaner maintenance. Perhaps
this will ultimately allow us to write gWasm compatible code for
Brass Golem without the necessecity of downloading and installing
Emscripten toolchain ourselves. Perhaps the tool could take care of
it for us!