Make gfaas decorated functions asynchronous
This commit makes `gfaas` decorated functions by default `async`
which essentially means they now return `impl Future` and can be
polled independently. In a way, now you can think of `gfaas`
decorated function as of `tokio::task::spawn_blocking` with this
difference that the body of the function is executed on Golem
rather than on the local machine.