Handle errors in user code in expanded function code
Previously, we'd panic in the expanded function code annotated with
`remote_fn`, which is less-than-ideal. Currently, the proposed approach
is to *always* return an `anyhow::Result` regardless of whether the user
expects it in their function signature or not. This implies that,
when awaiting the results, they will always have to handle/unwrap
the result *at least* once:

```rust
let res = hello("...").await?;
```

if `fn hello(String) -> String` for instance, or

```rust
let res = hello("...").await??;
```

if `fn hello(String) -> Result<String>`.

In other words, we always wrap the user expected return type in a
`Result` with errors from Yagna machinery, etc.