Simplify cross-compilation of functions to wasm32-wasi
Previously, gfaas-cli would force-install the wasm binaries. Now, instead
it builds them directly in the `gfaas_modules` project, and copies them
out so that they are relatively side-by-side with the main binary.
This should hopefully streamline the process of packaging the binary
with Wasm inserts in the future into the final app.